Multiple quality of life improvements - Remove usage of reflection - add .env to .gitignore - add support for extended resonance filter range in IT - clean up and fill out effect player interfaces - upgrade voice package to take advantage of more generics support - remove stupid usages of generics - split out structures from player interfaces

+++ internal/format/internal/filter/amigafilter.go
@@ -61,5 +61,5 @@ func (f *AmigaLPF) Filter(dry volume.Matrix) volume.Matrix {
 }
 
 // UpdateEnv updates the filter with the value from the filter envelope
-func (f *AmigaLPF) UpdateEnv(v float32) {
+func (f *AmigaLPF) UpdateEnv(v int8) {
 }
